+++ scripts/inventory.sh
@@ -1,4 +1,7 @@
 #!/usr/bin/env bash
+# SPDX-License-Identifier: GPL-3.0-only
+# Copyright (C) 2026 Manuel Alejandro Hernández Giuliani
+
 set -Eeuo pipefail
 
 umask 077